Ch05 Serialize out the player inventory from the store inventory, so the items persist once the player leaves the store
Added static methods to deal with ownership of items. Problems like putting items not purchased and exiting with them are now resolved.

When removing store items that were not purchased (upon exit of the store) the removals trigger a remove notification to update the status.

Buy button should be desensitized when it should

Stacked item removal from the player inventory when exiting from the store (without paying)

Added a method to deal with named inventory items (which we need when using the store inventory to differentiate ownership)

If you sold items and had enough money to buy an item that was already in your inventory (waiting to be purchased) the buy button should be enabled (a drop event)

When player items placed in store inventory, and then closing the store window. The issue is that the player items should know the index value in the players inventory. Created method to check to first available spots when saving.